requirements.txt
flask
SQLAlchemy
flask-sqlalchemy
pytz
python-dotenv
Flask-JWT-Extended
flask_cors
Flask-Caching
redis
celery[redis]
WeasyPrint
requests